Combine the 2 profile dialogs' ui
