+++ ZenovaLauncher/Dialogs/ProfileDialog.xaml
@@ -1,0 +1,134 @@
+﻿<ui:ContentDialog x:Class="ZenovaLauncher.ProfileDialog"
+      xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
+      xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
+      xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006" 
+      xmlns:d="http://schemas.microsoft.com/expression/blend/2008" 
+      xmlns:local="clr-namespace:ZenovaLauncher"
+      xmlns:ui="http://schemas.modernwpf.com/2019"
+      mc:Ignorable="d"
+      Style="{StaticResource LauncherContentDialog}"
+      HorizontalContentAlignment="Center"
+      SecondaryButtonText="Cancel"
+      CloseButtonText="Close"
+      DefaultButton="Primary">
+
+	<Grid>
+		<ui:SimpleStackPanel VerticalAlignment="Stretch" HorizontalAlignment="Center" Spacing="30" Margin="21" Width="400">
+			<ui:SimpleStackPanel VerticalAlignment="Stretch" HorizontalAlignment="Stretch" Spacing="2">
+				<TextBlock Margin="2,-8,0,0" FontSize="15" FontWeight="SemiBold" Typography.Capitals="AllSmallCaps" Text="Name" />
+				<TextBox 
+                    x:Name="ProfileNameBox"
+                    Style="{DynamicResource LauncherTextBoxStyle}"
+                    HorizontalAlignment="Stretch"
+					ui:ControlHelper.PlaceholderText="unnamed profile"/>
+			</ui:SimpleStackPanel>
+			<ui:SimpleStackPanel VerticalAlignment="Stretch" HorizontalAlignment="Stretch" Spacing="2">
+				<TextBlock Margin="2,-8,0,0" FontSize="15" FontWeight="SemiBold" Typography.Capitals="AllSmallCaps" Text="Version" />
+				<ComboBox
+                    x:Name="VersionBox"
+                    Style="{DynamicResource DialogComboBoxStyle}"
+                    SelectionChanged="VersionChanged"
+                    HorizontalAlignment="Stretch">
+					<ComboBox.ItemTemplate>
+						<DataTemplate>
+							<TextBlock Text="{Binding Path=FullName}" />
+						</DataTemplate>
+					</ComboBox.ItemTemplate>
+				</ComboBox>
+			</ui:SimpleStackPanel>
+			<Expander
+                x:Name="ModOptionsExpander"
+                ExpandDirection="Down"
+                IsExpanded="False"
+                VerticalAlignment="Top"
+                Padding="2">
+				<Expander.Header>
+					<TextBlock FontSize="15" FontWeight="SemiBold" Typography.Capitals="AllSmallCaps" VerticalAlignment="Center" Text="Mod Options" />
+				</Expander.Header>
+				<ui:SimpleStackPanel VerticalAlignment="Stretch" HorizontalAlignment="Stretch" Spacing="30" Margin="0,15,0,0">
+					<ui:SimpleStackPanel VerticalAlignment="Stretch" HorizontalAlignment="Stretch" Spacing="2">
+						<TextBlock Margin="2,-8,0,0" FontSize="15" FontWeight="SemiBold" Typography.Capitals="AllSmallCaps" Text="Available Mods" />
+						<Border Background="{DynamicResource LauncherControlTransientBackgroundBrush}" CornerRadius="2" Padding="0,2,0,2">
+							<ListBox
+                                x:Name="AvailableModsBox"
+                                Background="Transparent"
+                                Height="100"
+								ItemsSource="{Binding Path=LoadedMods}"
+                                ItemContainerStyle="{DynamicResource ProfileListBoxItemStyle}">
+								<ListBox.ItemTemplate>
+									<DataTemplate>
+										<Grid>
+											<Separator x:Name="Separator" Margin="12,0,12,0" VerticalAlignment="Top" Background="{DynamicResource AppBarSeparatorLowForeground}" />
+											<Grid Margin="12,3,12,3">
+												<TextBlock FontSize="14" FontWeight="DemiBold" Text="{Binding Name}" />
+												<Button x:Name="AddButton" HorizontalAlignment="Right" Padding="2" Visibility="Hidden" Click="AddModClick">
+													<Button.Content>
+														<ui:SimpleStackPanel Orientation="Horizontal" Spacing="5">
+															<TextBlock FontSize="12" Text="Add"/>
+															<ui:FontIcon Glyph="&#xECC8;" FontSize="12" />
+														</ui:SimpleStackPanel>
+													</Button.Content>
+												</Button>
+											</Grid>
+										</Grid>
+										<DataTemplate.Triggers>
+											<DataTrigger Binding="{Binding RelativeSource={RelativeSource AncestorType=ListBoxItem}, Path=IsMouseOver}" Value="True">
+												<Setter TargetName="AddButton" Property="Visibility" Value="Visible" />
+											</DataTrigger>
+											<DataTrigger Binding="{Binding RelativeSource={RelativeSource AncestorType=ListBoxItem}, Path=IsSelected}" Value="True">
+												<Setter TargetName="AddButton" Property="Visibility" Value="Visible" />
+											</DataTrigger>
+											<DataTrigger Binding="{Binding RelativeSource={RelativeSource PreviousData}}" Value="{x:Null}">
+												<Setter Property="Visibility" TargetName="Separator" Value="Collapsed"/>
+											</DataTrigger>
+										</DataTemplate.Triggers>
+									</DataTemplate>
+								</ListBox.ItemTemplate>
+							</ListBox>
+						</Border>
+					</ui:SimpleStackPanel>
+					<ui:SimpleStackPanel VerticalAlignment="Stretch" HorizontalAlignment="Stretch" Spacing="2">
+						<TextBlock Margin="2,-8,0,0" FontSize="15" FontWeight="SemiBold" Typography.Capitals="AllSmallCaps" Text="Loaded Mods" />
+						<Border Background="{DynamicResource LauncherControlTransientBackgroundBrush}" CornerRadius="2" Padding="0,2,0,2">
+							<ListBox
+                                x:Name="LoadedModsBox"
+                                Background="Transparent"
+                                Height="100"
+                                ItemContainerStyle="{DynamicResource ProfileListBoxItemStyle}">
+								<ListBox.ItemTemplate>
+									<DataTemplate>
+										<Grid>
+											<Separator x:Name="Separator" Margin="12,0,12,0" VerticalAlignment="Top" Background="{DynamicResource AppBarSeparatorLowForeground}" />
+											<Grid Margin="12,3,12,3">
+												<TextBlock FontSize="14" FontWeight="DemiBold" Text="{Binding Name}" />
+												<Button x:Name="RemoveButton" HorizontalAlignment="Right" Padding="2" Visibility="Hidden" Click="RemoveModClick">
+													<Button.Content>
+														<ui:SimpleStackPanel Orientation="Horizontal" Spacing="5">
+															<TextBlock FontSize="12" Text="Remove"/>
+															<ui:FontIcon Glyph="&#xECC9;" FontSize="12" />
+														</ui:SimpleStackPanel>
+													</Button.Content>
+												</Button>
+											</Grid>
+										</Grid>
+										<DataTemplate.Triggers>
+											<DataTrigger Binding="{Binding RelativeSource={RelativeSource AncestorType=ListBoxItem}, Path=IsMouseOver}" Value="True">
+												<Setter TargetName="RemoveButton" Property="Visibility" Value="Visible" />
+											</DataTrigger>
+											<DataTrigger Binding="{Binding RelativeSource={RelativeSource AncestorType=ListBoxItem}, Path=IsSelected}" Value="True">
+												<Setter TargetName="RemoveButton" Property="Visibility" Value="Visible" />
+											</DataTrigger>
+											<DataTrigger Binding="{Binding RelativeSource={RelativeSource PreviousData}}" Value="{x:Null}">
+												<Setter Property="Visibility" TargetName="Separator" Value="Collapsed"/>
+											</DataTrigger>
+										</DataTemplate.Triggers>
+									</DataTemplate>
+								</ListBox.ItemTemplate>
+							</ListBox>
+						</Border>
+					</ui:SimpleStackPanel>
+				</ui:SimpleStackPanel>
+			</Expander>
+		</ui:SimpleStackPanel>
+	</Grid>
+</ui:ContentDialog>
